% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{calculate_shannon_entropy}
\alias{calculate_shannon_entropy}
\title{Calculate Shannon Entropy}
\usage{
calculate_shannon_entropy(frequencies)
}
\arguments{
\item{frequencies}{Named numeric vector of codon frequencies}
}
\value{
Numeric value of Shannon entropy in bits
}
\description{
Calculate Shannon entropy of a codon frequency distribution.
}
\examples{
sequence <- "ATGATGATGTTATTATTACGCCGCCGCC"
freqs <- calculate_codon_frequencies(sequence)
entropy <- calculate_shannon_entropy(freqs)
}
