% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codon_frequency_df}
\alias{codon_frequency_df}
\title{Convert Codon Frequency Vector to Tidy Data Frame}
\usage{
codon_frequency_df(frequencies)
}
\arguments{
\item{frequencies}{Named numeric vector of codon frequencies}
}
\value{
Tibble with columns codon, frequency, aa
}
\description{
Convert Codon Frequency Vector to Tidy Data Frame
}
\examples{
freqs <- calculate_codon_frequencies("ATGATGATG")
codon_frequency_df(freqs)
}
