% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_all_codons}
\alias{.get_all_codons}
\title{Get All Possible Codons}
\usage{
.get_all_codons()
}
\value{
Character vector of 64 codons
}
\description{
Generate a vector of all 64 possible codons.
}
\keyword{internal}
