% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_codon_aa_map}
\alias{.get_codon_aa_map}
\title{Mapping amino acids to their corresponding codons}
\usage{
.get_codon_aa_map()
}
\value{
Corresponing amino acid\(s\)
}
\description{
Helper function to add amino acid to codon.
}
\keyword{internal}
