% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_window_coords}
\alias{.get_window_coords}
\title{Sliding Window Iterator}
\usage{
.get_window_coords(total_length, window_size, step_size)
}
\arguments{
\item{total_length}{Integer length of sequence}

\item{window_size}{Integer size of sliding window}

\item{step_size}{Integer step size for sliding window}
}
\value{
Data frame with window coordinates
}
\description{
Helper function to generate sliding window coordinates.
}
\keyword{internal}
