% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parse_gtf_line}
\alias{.parse_gtf_line}
\title{Parse GTF Line}
\usage{
.parse_gtf_line(line)
}
\arguments{
\item{line}{Character string of a GTF line}
}
\value{
List of parsed attributes
}
\description{
Helper function to parse a single GTF line.
}
\keyword{internal}
