% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_dna}
\alias{.validate_dna}
\title{Validate DNA Sequence}
\usage{
.validate_dna(sequence)
}
\arguments{
\item{sequence}{Character string of DNA sequence}
}
\value{
Boolean indicating if sequence is valid
}
\description{
Check if a DNA sequence contains only valid nucleotides.
}
\keyword{internal}
