% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{GeneScout-extdata}
\alias{GeneScout-extdata}
\title{Example FASTA and GTF files bundled with GeneScout}
\description{
GeneScout includes a collection of small to moderately sized FASTA and GTF
files intended for examples, testing, and vignettes. These files are stored
in compressed form under \code{inst/extdata} to reduce package size and
improve portability.
}
\details{
The datasets include:
\itemize{
\item Human protein-coding and lncRNA transcript FASTA files from GENCODE v19
\item Human chromosome 19 genomic sequence (GRCh37) [\verb{REMOVED CURRENTLY!}]
\item Example gene sequences (e.g., \code{APOE}, \code{TP53})
\item A nematode genome and gene annotation (Acanthocheilonema viteae)
}

Files can be accessed using \code{\link[base]{system.file}}:

\preformatted{
system.file("extdata", "APOE.fasta.gz", package = "GeneScout")
}

All files are provided strictly for demonstration and benchmarking purposes
and should not be considered authoritative reference genomes.
}
\keyword{datasets}
