% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{translate_codons}
\alias{translate_codons}
\title{Translate Codons to Amino Acids}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{Character vector of codons (e.g. "ATG", "GCC")}
}
\value{
Character vector of amino acids (single-letter codes, "*" for stop)
}
\description{
Translate DNA codons into single-letter amino acid codes
using the standard genetic code.
}
\examples{
translate_codons(c("ATG", "TAA", "GCC"))
}
